move ssh form interface from model to action
Feature #14479_ssh_keys

Arvados-DCO-1.1-Signed-off-by: Janicki Artur <artur.janicki@contractors.roche.com>
