+++ src/views-components/dialog-forms/create-ssh-key-dialog.ts
@@ -5,9 +5,8 @@
 import { compose } from "redux";
 import { reduxForm } from 'redux-form';
 import { withDialog } from "~/store/dialog/with-dialog";
-import { SSH_KEY_CREATE_FORM_NAME, createSshKey } from '~/store/auth/auth-action';
+import { SSH_KEY_CREATE_FORM_NAME, createSshKey, SshKeyCreateFormDialogData } from '~/store/auth/auth-action';
 import { DialogSshKeyCreate } from '~/views-components/dialog-create/dialog-ssh-key-create';
-import { SshKeyCreateFormDialogData } from '~/models/ssh-key';
 
 export const CreateSshKeyDialog = compose(
     withDialog(SSH_KEY_CREATE_FORM_NAME),
